Predict parameter values from a fitted model

This method generates predicted parameter values for the original dataset or a new dataset using the posterior distribution. Standard deviation and a customizable uncertainty interval are also calculated.

# S4 method for ubmsFit
predict(
  object,
  submodel,
  newdata = NULL,
  transform = TRUE,
  re.form = NULL,
  level = 0.95,
  ...
)

Arguments

object
A fitted model of class ubmsFit

submodel
Submodel to predict from, for example "det"

newdata
Optional data frame or RasterStack of covariates to generate predictions from. If not provided (the default), predictions are generated from the original data

transform
If TRUE, back-transform the predictions to their original scale

re.form
If NULL, any estimated group-level parameters ("random effects") are included. If NA, they are ignored

level
Probability mass to include in the uncertainty interval

...
Currently ignored

Value

If newdata was a data frame: A data frame with one row per prediction and four columns: 1) Predicted point estimates (posterior means), 2) Standard deviation of the posterior, 3-4) Lower and upper bounds of the specified uncertainty interval

For parameters with more than one dimension, the rows are in site-major order, or site-year-observation for dynamic models.

If newdata was a RasterStack, returns a RasterStack with four layers corresponding to the four columns above with the same projection as the original RasterStack.

See also