Fit the Occupancy Model of Royle and Nichols (2003)

Fit the occupancy model of Royle and Nichols (2003), which relates probability of detection of the species to the number of individuals available for detection at each site.

stan_occuRN(
  formula,
  data,
  K = 20,
  prior_intercept_state = normal(0, 5),
  prior_coef_state = normal(0, 2.5),
  prior_intercept_det = logistic(0, 1),
  prior_coef_det = logistic(0, 1),
  prior_sigma = gamma(1, 1),
  ...
)

Arguments

formula
Double right-hand side formula describing covariates of detection and abundance in that order

data
A unmarkedFrameOccu object

K
Integer upper index of integration for N-mixture. This should be set high enough so that it does not affect the parameter estimates. Note that computation time will increase with K.

prior_intercept_state
Prior distribution for the intercept of the state (abundance) model; see ?priors for options

prior_coef_state
Prior distribution for the regression coefficients of the state model

prior_intercept_det
Prior distribution for the intercept of the detection probability model

prior_coef_det
Prior distribution for the regression coefficients of the detection model

prior_sigma
Prior distribution on random effect standard deviations

...
Arguments passed to the stan call, such as number of chains chains or iterations iter

Value

ubmsFitOccuRN object describing the model fit.

References

Royle JA, Nichols JD. 2003. Estimating abundance from repeated presence-absence data or point counts. Ecology 84: 777-790.

See also

Examples

# \donttest{
data(birds)
woodthrushUMF <- unmarkedFrameOccu(woodthrush.bin)
#Add a site covariate
siteCovs(woodthrushUMF) <- data.frame(cov1=rnorm(numSites(woodthrushUMF)))

(fm_wood <- stan_occuRN(~1~cov1, woodthrushUMF, chains=3, iter=300))

#> 
#> SAMPLING FOR MODEL 'occuRN' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.002413 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 24.13 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 300 [  0%]  (Warmup)
#> Chain 1: Iteration:  30 / 300 [ 10%]  (Warmup)
#> Chain 1: Iteration:  60 / 300 [ 20%]  (Warmup)
#> Chain 1: Iteration:  90 / 300 [ 30%]  (Warmup)
#> Chain 1: Iteration: 120 / 300 [ 40%]  (Warmup)
#> Chain 1: Iteration: 150 / 300 [ 50%]  (Warmup)
#> Chain 1: Iteration: 151 / 300 [ 50%]  (Sampling)
#> Chain 1: Iteration: 180 / 300 [ 60%]  (Sampling)
#> Chain 1: Iteration: 210 / 300 [ 70%]  (Sampling)
#> Chain 1: Iteration: 240 / 300 [ 80%]  (Sampling)
#> Chain 1: Iteration: 270 / 300 [ 90%]  (Sampling)
#> Chain 1: Iteration: 300 / 300 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 3.22519 seconds (Warm-up)
#> Chain 1:                3.02242 seconds (Sampling)
#> Chain 1:                6.24761 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'occuRN' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.001932 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 19.32 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:   1 / 300 [  0%]  (Warmup)
#> Chain 2: Iteration:  30 / 300 [ 10%]  (Warmup)
#> Chain 2: Iteration:  60 / 300 [ 20%]  (Warmup)
#> Chain 2: Iteration:  90 / 300 [ 30%]  (Warmup)
#> Chain 2: Iteration: 120 / 300 [ 40%]  (Warmup)
#> Chain 2: Iteration: 150 / 300 [ 50%]  (Warmup)
#> Chain 2: Iteration: 151 / 300 [ 50%]  (Sampling)
#> Chain 2: Iteration: 180 / 300 [ 60%]  (Sampling)
#> Chain 2: Iteration: 210 / 300 [ 70%]  (Sampling)
#> Chain 2: Iteration: 240 / 300 [ 80%]  (Sampling)
#> Chain 2: Iteration: 270 / 300 [ 90%]  (Sampling)
#> Chain 2: Iteration: 300 / 300 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.25745 seconds (Warm-up)
#> Chain 2:                2.5902 seconds (Sampling)
#> Chain 2:                4.84765 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'occuRN' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.001895 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 18.95 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:   1 / 300 [  0%]  (Warmup)
#> Chain 3: Iteration:  30 / 300 [ 10%]  (Warmup)
#> Chain 3: Iteration:  60 / 300 [ 20%]  (Warmup)
#> Chain 3: Iteration:  90 / 300 [ 30%]  (Warmup)
#> Chain 3: Iteration: 120 / 300 [ 40%]  (Warmup)
#> Chain 3: Iteration: 150 / 300 [ 50%]  (Warmup)
#> Chain 3: Iteration: 151 / 300 [ 50%]  (Sampling)
#> Chain 3: Iteration: 180 / 300 [ 60%]  (Sampling)
#> Chain 3: Iteration: 210 / 300 [ 70%]  (Sampling)
#> Chain 3: Iteration: 240 / 300 [ 80%]  (Sampling)
#> Chain 3: Iteration: 270 / 300 [ 90%]  (Sampling)
#> Chain 3: Iteration: 300 / 300 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.53477 seconds (Warm-up)
#> Chain 3:                2.59534 seconds (Sampling)
#> Chain 3:                5.13011 seconds (Total)
#> Chain 3: 

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Warning: Some Pareto k diagnostic values are slightly high. See help('pareto-k-diagnostic') for details.

#> 
#> Call:
#> stan_occuRN(formula = ~1 ~ cov1, data = woodthrushUMF, chains = 3, 
#>     iter = 300)
#> 
#> Abundance (log-scale):
#>             Estimate    SD   2.5% 97.5% n_eff  Rhat
#> (Intercept)   0.7663 0.171  0.435 1.097   169 1.021
#> cov1          0.0845 0.124 -0.169 0.324   222 0.998
#> 
#> Detection (logit-scale):
#>  Estimate    SD  2.5%  97.5% n_eff Rhat
#>      -1.2 0.174 -1.57 -0.883   244 1.02
#> 
#> LOOIC: 638.539
#> Runtime: 16.225 sec

# }